move dataclasses to energy_source
